[O] Strip the easter egg from Profile.vue requested by @hykilpikonna
